app/requirements.txt
Django==4.2.16
django-environ==0.11.2
djangorestframework==3.15.2
psycopg2==2.9.9
django-cors-headers==4.4.0
uuid6==2024.7.10
celery==5.4.0
gunicorn==23.0.0
requests==2.32.3
drf_social_oauth2==3.1.0
openai==1.55.3

# 支払いについて
stripe==12.0.1

# dev
ruff==0.8.1
PyYAML==6.0.2
## セキュリティチェック
bandit==1.8.0
## 型チェック
mypy==1.13.0
django-stubs[compatible-mypy]==5.1.1
